s053: shape probes 2/3 — ceiling is the FRONT END not the leaf (MLP 2x1024 on s052 feats 0.73 fresh/0.23 held; joint shape x fill supervision no help); fields 0.88 vs geometric 3-way 0.58; boundary-orientation block 92-d = 900-d front end (0.709 vs 0.705), +colour 192-d 0.761, held-out 0.23->0.35, small 0.37->0.46; corner block +1pp only (fills make interior corners); feature cache; CNN dataset-ceiling ref (running)

+++ experiments/progenitor/shapes.py
@@ -101,7 +101,7 @@ def sample(n, seed, *, maxk=1, p_iso=0.15, p_crop=0.2, exclude=(), only=None, no
     pairs = [(s, f) for s in shape_choices for f in (fill_choices if s < 3 else (0,))]
     if only is not None: pairs = [p for p in pairs if p in set(only)]
     pairs = [p for p in pairs if p not in set(exclude)]
-    pb = torch.tensor(p_blur); pf = torch.tensor(p_focus)
+    pb = torch.tensor(p_blur, dtype=torch.float); pf = torch.tensor(p_focus, dtype=torch.float)
     for i in range(n):
         hb, sb, vb = R(), R() * 0.6, R() * 0.6 + 0.3; bg = _hsv(hb, sb, vb)
         k = int(torch.randint(1, maxk + 1, (1,), generator=gen)) if maxk > 1 else 1
